+++ src/core/components/documentParser/documentParser.tsx
@@ -1,4 +1,5 @@
 import { Image } from '@tiptap/extension-image';
+import TipTapLink from '@tiptap/extension-link';
 import { EditorContent, useEditor } from '@tiptap/react';
 import { StarterKit } from '@tiptap/starter-kit';
 import classNames from 'classnames';
@@ -6,30 +7,56 @@ import { useEffect, type ComponentProps } from 'react';
 import sanitizeHtml from 'sanitize-html';
 import { Markdown } from 'tiptap-markdown';
 
-interface IDocumentParserProps extends Omit<ComponentProps<'div'>, 'ref'> {
-    children: string;
+export interface IDocumentParserProps extends Omit<ComponentProps<'div'>, 'ref'> {
+    stringDocument: string;
 }
 
 export const DocumentParser: React.FC<IDocumentParserProps> = (props) => {
-    const { children, className, ...otherProps } = props;
+    const { children, className, stringDocument, ...otherProps } = props;
     const parser = useEditor({
         editable: false,
-        extensions: [StarterKit, Markdown, Image],
+        extensions: [
+            StarterKit.configure({
+                codeBlock: {
+                    HTMLAttributes: {
+                        class: 'language-html',
+                    },
+                },
+            }),
+            Image,
+            Markdown,
+            TipTapLink.configure({
+                openOnClick: 'whenNotEditable',
+            }),
+        ],
     });
 
     useEffect(() => {
         if (parser) {
-            const safeHTML = sanitizeHtml(children, {
+            const safeHTML = sanitizeHtml(stringDocument, {
+                allowedTags: sanitizeHtml.defaults.allowedTags.concat([
+                    'img',
+                    'del',
+                    'video',
+                    'audio',
+                    'svg',
+                    'code',
+                    'pre',
+                ]),
+                allowedClasses: {
+                    code: ['language-*', 'lang-*'],
+                },
                 allowedAttributes: {
                     ...sanitizeHtml.defaults.allowedAttributes,
                     img: ['src', 'alt'],
                     a: ['href', 'title'],
                 },
+
                 disallowedTagsMode: 'recursiveEscape',
             });
             parser.commands.setContent(safeHTML, true);
         }
-    }, [parser, children]);
+    }, [parser, stringDocument]);
 
     const proseClassnames = classNames(
         'prose', // enable prose
@@ -57,5 +84,5 @@ export const DocumentParser: React.FC<IDocumentParserProps> = (props) => {
         className,
     );
 
-    return <EditorContent editor={parser} className={proseClassnames} {...otherProps} />;
+    return <EditorContent editor={parser} className={proseClassnames} data-testid="doc-parser" {...otherProps} />;
 };