+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `CopyParser` and `Accordion` core component
+-   Implement `DocumentParser` and `Accordion` core component
 
 ## [1.0.25] - 2024-04-24